requirements.txt
aiogram==2.25.1
gspread==5.0.0
google-auth==2.0.0
